Split lib.rs
